CE-2785 Register hooks in hooks class
